add(junyeong/todoComment/#15): todo일정 삭제시 todo댓글 삭제 연동 구현, delete와 edit시 성공여부 확인 구현

+++ app/src/models/todo/todoCommentStorage.js
@@ -20,29 +20,43 @@ class TodoCommentStorage {
   static async addComment(client) {
     const sql =
       "INSERT INTO todo_comment (user_no, writer_no, date, content) VALUES (?,?,?,?)";
-    console.log(client.user_id);
     const user_no = await DataCheck.getUserNo(client.user_id);
     const writer_no = await DataCheck.getUserNo(client.writer_id);
     const req = [user_no, writer_no, client.date, client.content];
-    const result = await db.query(sql, req);
+    const result = { success: false };
+    const isCheck = (await db.query(sql, req))[0].affectedRows;
+    if (isCheck) result.success = true;
 
-    return result[0];
+    return result;
   }
 
   static async editComment(client) {
     const sql = "UPDATE todo_comment SET content= ? WHERE no= ?;";
     const req = [client.content, client.no];
-    const result = await db.query(sql, req);
+    const result = { success: false };
+    const isCheck = (await db.query(sql, req))[0].affectedRows;
+    if (isCheck) result.success = true;
 
-    return result[0];
+    return result;
   }
 
   static async deleteComment(client) {
-    const sql = "DELETE FROM todo_comment WHERE no= ?";
-    const req = [client.id];
-    const result = await db.query(sql, req);
+    let sql, req;
+    const result = { success: false };
+    if (client === 0) {
+      return (result.success = true);
+    }
+    if (!client.id) {
+      sql = "DELETE FROM todo_comment WHERE date= ?";
+      req = client;
+    } else {
+      sql = "DELETE FROM todo_comment WHERE no= ?";
+      req = client.id;
+    }
+    const isCheck = (await db.query(sql, req))[0].affectedRows;
+    if (isCheck) result.success = true;
 
-    return result[0];
+    return result;
   }
 }
